refactor: renamed and refactored 001_hard_delete_old.py
